+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#362](https://github.com/babylonlabs-io/vigilante/pull/362) chore: manage overlap block range
 * [#369](https://github.com/babylonlabs-io/vigilante/pull/369) chore: do not reprocess block heights - but for real now
 * [#370](https://github.com/babylonlabs-io/vigilante/pull/370) chore: staking event watcher add retry
+* [#376](https://github.com/babylonlabs-io/vigilante/pull/376) chore: bump babylon node dep to v2
 
 ### Bug Fixes
 * [#359](https://github.com/babylonlabs-io/vigilante/pull/359) fix: correctly increment page for stk hash query